requirements.txt

discord.py
easyocr
Pillow
